Bugfix: Correct spelling mistake, change icon, reduce icon text length
There does happen to be a more suitable icon for the full screen view toolbar
button in the Oxygen icon set.  It is only used for Mudlet on a small screensize
display, this commit adds it to Mudlet whilst also contracting the text
under the icon to be a bit smaller so the overal button takes up less space.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>