BugFix/Enhancement: Custom Line context menu/GUI
The ability to draw new custom exit lines had been lost in the previous
SF git master.  This reverts the GUI changes previously made and enhances
the results.  The former dialog is now more intuitive.  It is now possible
to (repetitively) undo line segments placed during the drawing operation
and to modify the characteristics of the line both during drawing and
subsequently.  It is also now possible to modify an existing line by adding
or removing segments also via an enhanced context menu.

An Icon for the new dialog (custom_lines_properties.ui) has been added to
the resources file and icons also created for custom_lines.ui and
room_exits.ui.  In doing this it became clear that the icon used for the
main window was the low resolution one (16x16) and the (48x48) looks better.
Also the About [mudlet] action on the main toolbar was using a mudlet ones
but there is a different one that is normally used for this action.
a widely used standard one that

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>