Enhancement: Adds control of Mudlet Version to project file.
The splash screen is changed so that the version text has been removed
from the background image, it is now shown superimposed at the centre
top of the image as the loading text message grows downwards during
the loading process.  The version text will explicitly say that it
is a development version unless the BUILD value in the project file
is blank.

Adds a formatted copy of version 3 of the GPL to supplement just
the previous reference to it on the About dialog "Licence" tab.

Also adds me to named contributors on the About dialog!

Adds a getVersion() LUA function so that future script developers
can tailor their work to accommodate new features and enhancements
if available on the user's client.  As the build "suffix" is also
made available, if the BUILD variable (APP_BUILD in source code) in
the project file is not left blank then scripts can also be made
aware of custom or developer's versions.  As uploaded the string was
left as "-rc2-slysven" for release candidate 2, sourced from my
repository instead of main one, I urge other developers to adjust this
for the code in their own repositories... 8-)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>