Enhancement: Add --help / --version command line arguments (FSF compliance)
The Free Software Federation strongly urges a pair of standard command line
arguments and the responses they should invoke.  Also the GPL v3 indicated
the sort of information that a program with a graphical user inter should
place on its splash screen.  This change aims to provide this, as well the
--help / -h option also documents the command line arguments that the Qt
libraries provide automagically, though they do not seem particularly useful
at this time.  It is possible that there is a further option that I recall
seeing somewhere but can not find that had a use for debugging.

Also implemented is a -q / --quiet option that suppresses the splash screen
entirely, to alert the user to the start-up delay the cursor is changed to
the waiting one during the startup whether the splash screen is suppressed
or not.

The Windows OS should respond to "/h" (or "/?") and "/v" command line switches
though whether the Qt ones are recognised in this format is not clear at this
point, needs testing by a user of such an OS.

A correction has been made to the GPL licence text available on the "Licence"
tab of the "About" dialog, to restore a couple of missing paragraphs omitted
from the end of the text.  The introduction has been corrected so that it
indicated the Licence is "GPL2+" which matches the text in the source files
as opposed to the "GPL3" statement originally on that specific dialog which
contradicted the rest of the application files.  The Mudlet version text has
been move to be superimposed over the Mudlet logo - along with a copyright
text so that it mimics the layout from the same graphic as used on the splash
screen.