Enhancement: Debug controls framework for developer run-time modifications
Provide places (6) to put in code elements that place controls into the
last tab in the dlgProfilePreferences.cpp dialog that can be used to modify,
possibly temporarily, the operation of the program on a per host basis.
Might be used to allow the run-time selection of alternate pieces of
code for performance comparisons or to investigate suspect code.  As is
the changes are not saved between runs of the application.

Included here is a spin-box control to reduce the alpha of the colour
used to draw the rooms on the 2D map, which I needed to check the placement
of exits lines close to/under the room symbol.

The extra lines in the comments are intentional to allow sufficient space
between individual controls so that patching to insert or remove them
can be done automagically (the patch program has an easier time of it
if lines around changes remain stable through changes- or so I understand).

N.B. Host.cpp gains the Q_OBJECT macro so qmake WILL need to be run to ensure
it gets run though the "moc" before a make is done!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>