Enhancement: Debug controls framework for developer run-time modifications
Provide places (10) to put in code elements that place controls into the
last tab in the dlgProfilePreferences.cpp dialog that can be used to modify,
possibly temporarily, the operation of the program on a per host basis or
application wide.  Useful to allow the run-time selection of alternate
pieces of code for performance comparisons or to investigate suspect code.
As is the changes are not saved between runs of the application.

Adds a missing icon to the Profile Preferences dialog.

Included here is a spin-box control to reduce the alpha of the colour
used to draw the rooms on the 2D map, which I needed to check the placement
of exits lines close to/under the room symbol.

The extra lines in the comments are intentional to allow sufficient space
between individual controls so that patching to insert or remove them
can be done automagically (the patch program has an easier time of it
if lines around changes remain stable through changes- or so I understand).

N.B. Host.cpp gains the Q_OBJECT macro so qmake WILL need to be run to
ensure it gets run though the "moc" when the make is done!

If the Todo plugin is available and enabled in QtCreator IDE one can locate
the places where the code is by adding the string "DEBUGCONTROLS" as a
title of one of things to track as this and a following spaced digit '0'
to '9' is embedded in the opening comment for each of the placeholders for
blocks of code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>