Revise: Splash-screen code now takes long ver. text, copied to About dialog
The previous code to draw copyright and version text onto the Splash screen
image truncated version texts that were over long (as might be the case for
where the qmake project file BUILD variable was set to a long, multiple
word one).  This revision shrinks the font size used for the version text
until it can be drawn on one line.

For uniformity and for the same reason, the same code is also copied across
to where the same image is used on the About Mudlet dialogue.

This commit was Cherry-picked from a number of pending revisions but is
believed to completely described the required changes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>