Tweak: missing file to project file, omit 2 connects with missing slots
A Ui file custom_lines_properties.ui got missed out of project file,
though application worked without the mention as it is carried inside
the resource file.  Also added a couple of blank lines and set project
build variable to something appropriate *for me*.

Two connect() usages in the dlgProfilePreferences.cpp file refer to SLOTs
that have NEVER existed in either the header file (as a declaration) or
in the source file (as a definition).  The lines concerned date back many
years and the possible functionality of one of them might be a subset of
a Lua command.  Without the implementation of the SLOTs all that happens
is a few lines of warning messages each time the Profile Preferences
dialog is called up.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>