BugFix/Enhance: Fix TELNET TTYPE entry, enhance behaviour ready for MTTS
Telnet Sub-option 24 (TERMINAL TYPE) only allows A-Z,a-z,0-0,- and /
characters, is a maximum of 40 character and does NOT seem to allow
spaces.  A previous change permitted setting of the version data in the
project file which makes its way into the TTYPE entry but did not address
these restrictions.

A feature of this particular option is that a Telnet client may respond
more than once, and for responses that describe devices that are
similar the most detailed response should be returned first, this commit
implements this in such a way that if the version details include a build
version the first response is of the form "MUDLET-X-Y-Z-B" where X, Y & Z
are the Major, Minor, and Revision version numbers and B is the
(Uppercase) build version, truncated to a maximum of 40 characters.  The
next (or first for a "Release" build with no Build version) is
"MUDLET-X-Y-Z" and after that just "MUDLET".  A further SubOption request
returns that last value a second time which is the signal to the
enquiring Server that it IS the last value in the list of response.

For a client that returns multiple responses that are functionally
different the client is expect to adopt the characteristics of the last
response it gives, so that the Server can select the type of terminal it
wishes to connect to.  In the case of Mudlet, because all the responses
describe the same level of functionality it does not need to react to
the choice a Server makes.

RFC1091 describes a mechanism for the server to return to the start of
the list by a further request which is expect to cycle the Client to the
start of the list - which is what happens.

There is an additional MUD specific extension, Mud Terminal Type
Specification (MTTS) which requires the last, repeated response to be
MTTS plus a number made from ASCII digit characters, this number being
the sum of a number of Mud Client characteristics each encoded with a
different power of 2 value.  Unfortunately there are some details of the
specification (at http://tintin.sourceforge.net/mtts/) that I am trying
to clarify with the author which seem to contradict RFC1091.  Once these
have been satisfactorily dealt with I will add the required code to do it.

I have also made some changes to make more of the debug output dependant
on the DEBUG_TELNET define, and to make the output from such a bit more
uniform in presentation and layout.  There were also some defined but
unused boolean flags that have been commented out.  Some DEFINES have been
renamed to have more uniform Names and a couple of duplicates eliminated.

I have also added a method that returns a mnemonic code for all telnet
sub-options that I have been able to identify the number for, which may
assist in identifying unexpected additions to a Server's protocol
handlers.

Sadly the indentation of cTelnet.cpp is not uniform in some places but
correcting it in this commit would produce too many changes to spot the
functional ones, a separate commit to follow will address that.